pim-compiler/output/react_agent_v3/user_management/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib
motor
pytest
httpx
email-validator